Add function to call psql.
This method differs from the psql.call_psql method in that it provides
sql statements via standard input rather than a file. When providing
sql via a file, psql provides helpful error messages that include the
full path of the file. In our case, we want our error messages to be
platform agnostic. Providing the sql via standard input works around
this issue.

Furthermore, we want to interleave stdout and stderr for our
tests. Thus we assign stderr to stdout when running the sql.